♻️ refactor(api/use-cases): separate email use-cases by application logic
