--- apps/api/src/application/use-cases/email/interface/email.usecase.interface.ts
@@ -1,5 +1,0 @@
-import type { CreateEmailOptions, CreateEmailRequestOptions, CreateEmailResponse } from "resend";
-
-export interface IEmailUseCase {
-	sendEmail(payload: CreateEmailOptions, options: CreateEmailRequestOptions): Promise<CreateEmailResponse>;
-}
