+++ apps/api/src/application/use-cases/email/interface/send-email.usecase.interface.ts
@@ -1,0 +1,5 @@
+import type { CreateEmailOptions, CreateEmailRequestOptions, CreateEmailResponse } from "resend";
+
+export interface ISendEmailUseCase {
+	execute(payload: CreateEmailOptions, options: CreateEmailRequestOptions): Promise<CreateEmailResponse>;
+}
