+++ apps/api/src/application/use-cases/email/type.ts
